Add heart rate mapping
